Выполняет задание "Функции возвращаются".
