+++ index.html
@@ -227,8 +227,8 @@
     </section>
   </template>
 
+  <script src="./js/main.js" type="module"></script>
   <script src="./js/functions.js"></script>
-  <script src="./js/main.js"></script>
 
 
 </body>
