+++ js/util.js
@@ -1,0 +1,9 @@
+// Функция для поиска случайного числа между а и b;
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export{getRandomInteger};
